+++ src/python/grapl_analyzerlib/grapl_analyzerlib/provision/queries.py
@@ -1,0 +1,47 @@
+from __future__ import annotations
+import json
+
+from typing import Mapping, TYPE_CHECKING, Any, Dict, List, cast
+
+from grapl_analyzerlib.grapl_client import GraphClient
+
+
+def query_dgraph_type(client: GraphClient, type_name: str) -> List[QueryPredicateResult]:
+    query = f"""
+        schema(type: {type_name}) {{ type }}
+    """
+    txn = client.txn(read_only=True)
+    try:
+        res = json.loads(txn.query(query).json)
+    finally:
+        txn.discard()
+
+    if not res:
+        return []
+    if not res.get("types"):
+        return []
+
+    res = res["types"][0]["fields"]
+    predicate_names = []
+    for pred in res:
+        predicate_names.append(pred["name"])
+
+    predicate_metas = []
+    for predicate_name in predicate_names:
+        predicate_metas.append(query_dgraph_predicate(client, predicate_name))
+
+    return predicate_metas
+
+
+QueryPredicateResult = Mapping[str, Any]
+def query_dgraph_predicate(client: GraphClient, predicate_name: str) -> Any:
+    query = f"""
+        schema(pred: {predicate_name}) {{  }}
+    """
+    txn = client.txn(read_only=True)
+    try:
+        res = json.loads(txn.query(query).json)["schema"][0]
+    finally:
+        txn.discard()
+
+    return res
